refactor(imports): update import paths for Instrument
